+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.7-2019.12.02
 * feature: edit rating for a song
+* feature: show album art in background
 
 ## 1.6-2019.12.02
 * feature: smart playlists